Tested ECS Performance
There seems to be a hard 10,000 limit on our current sprite renderer, but I doubt we'll reach that so I'm not too worried about it just yet. Similarly, I tested to make sure our ECS system wasn't slowing down sprite rendering; I did one batch of rendering with it and one without, and they performed roughly equally, so that assuages my fears that the ECS was burdening the game drastically. I still want to rewrite the collision code to be more performant though.