Preemptively Fixed a Quad Batching Bug
This code needs some cleaning-up, but it does fix a bug we would've run into sooner or later where a quad's map and its texture would've been split between two batches. Now, when we send a quad to OpenGL, we first ensure that the texture and map are always in the same batch.